Update Book Appointment.html
I have edited this code so that it can minimise without affecting the contents of the page, I'll still working on a few more things when it comes to minimising but so far, it's been working well. I am also going to upload a styles3.css file where I'll be updating more as time goes.